% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_est.R
\name{freq_est}
\alias{freq_est}
\title{Frequency Estimation Using the Optimized Period Technique}
\usage{
freq_est(x, delta_t = 1, f_min, f_max, names = TRUE)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{f_min}{a real value, the minimum frequency of the tested frequency range}

\item{f_max}{a real value, the maximum frequency of the tested frequency range}
}
\value{
the best-fitting frequency estimated with the XXX algorithm (see ***)
}
\description{
Frequency Estimation Using the Optimized Period Technique
}
\examples{
x1 = sin(0.11*2*pi*(1:100));
x2 = sin(0.11*2*pi*(1:100))+ 1.01*sin(0.17*2*pi*(1:100));
freq_est(x1)
freq_est(x2)

}
