% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_ICD.R
\name{freq_ICD}
\alias{freq_ICD}
\title{Frequency Estimation Using the Interval Length between Succeeding Maxima as Identified by an Algorithm Implemented in ICDs}
\usage{
freq_ICD(
  x,
  thresh_min = 0.02,
  amp_fac = 0.4,
  delta_t_u = 0.026,
  delta_t_l = 0.026,
  delta_t = 0.001,
  ...
)
}
\arguments{
\item{x}{a real valued array containing the values of the time series}

\item{thresh_min}{a positive real value, the length of the minimum threshold for peak detection}

\item{amp_fac}{a positive real value, the factor defining the upper threshold}

\item{delta_t_u}{a positive real value, the length of the upper threshold time interval (in time units), default value is delta_t_u = 0.026}

\item{delta_t_l}{a positive real value, the length of the lower threshold time interval (in time units), default value is delta_t_l = 0.026}

\item{delta_t}{a positive real value, the sampling time of the algorithm (in time units), default value is delta_t = 0.001}

\item{...}{optional arguments to be passed to the function splus2R::peaks. Otherwise the values span = 11, strict = FALSE and endbehavior = 0 are used.}
}
\value{
a named list containing the following components:
\itemize{
\item \code{freq_ICD} - the estimated frequency.
\item \code{c_v} - the coefficient of variation of the time intervals between  the identified peaks.
\item \code{no_of_peaks} - the number of peaks.
\item \code{maxima_sampled} -  the time series indices of the identified peaks.
}
}
\description{
This function implements the sensing algorithm used by defibrillators of the Biotronic company for estimating the
heart rate and the fibrillation frequency. The parameters are optimized for the data set included in this library, i.e., for mice ECGs.
}
\examples{
# Let's Consider two synthetic time series and an example from the attached data set:
x1 = sin(0.02 * 2*pi * (1:1000))
x2 = sin(0.02 *  2*pi * (1:1000)) + 1.1 * sin( 0.025 * 2*pi * (1:1000))
ecg_6  <- MiceFibECGs[,6]

freq_ICD(x1)
freq_ICD(x2)
# The irregularity of the fibrillation is indicated by a high value of the coefficient of variation
freq_ICD(ecg_6)
}
\references{
{ Brueggemann T, Dahlke D, Chebbo A, and Neumann I: \emph{Tachycardia detection in modern implantable cardioverter–defibrillars}. Herzschrittmachertherapie and Elektrophysiologie 27, 171-185 (2016) }

{ Diaz-Maue L, Witt A, Nobach H: \emph{Unraveling Cardiac Arrhythmia Frequency, Comparative Analysis Using Time and Frequency Domain Algorithms}. Front. Signal Process., Sec. Biomedical Signal Processing 5 (2025),  \url{https://doi.org/10.3389/frsip.2025.1707422}}‚
}
