% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lorenz_fit_periodogram.R
\name{Lorentz_fit_periodogram}
\alias{Lorentz_fit_periodogram}
\title{Frequency Estimation Using a Lorentz Fit to the Periodogram}
\usage{
Lorentz_fit_periodogram(
  x,
  delta_t = 1,
  welch_window = FALSE,
  f_min = 0,
  f_max = Inf
)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{welch_window}{if TRUE, the time series will be multiplied with a Welch window before Fourier transform}
}
\value{
a list containing the parameters of the Lorentz fit.
}
\description{
This function is fitting a Lorentz function to the periodogram of the time series x.
}
\examples{
x1 = sin(0.11*2*pi*(1:100));
x2 = sin(0.11*2*pi*(1:100))+ 1.01*sin(0.17*2*pi*(1:100));
x3 = sin(0.10*2*pi*(1:100))+ sin(0.20*2*pi*(1:100));
Lorentz_fit_periodogram(x1)
Lorentz_fit_periodogram(x2)
Lorentz_fit_periodogram(x2,  welch_window = TRUE)
Lorentz_fit_periodogram(x3)
}
