% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_adapted_variable_period.R
\name{freq_adapted_variable_period}
\alias{freq_adapted_variable_period}
\title{Frequency Estimation Using the Adapted Variable Period Technique}
\usage{
freq_adapted_variable_period(x, delta_t = 1)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}
}
\value{
a named list containing the following components:
\itemize{
\item \code{freq_AVP} - the best-fitting frequency estimated with the Adapted Variable Period Technique.
\item \code{Max_P} - the corresponding maximum value of the adapted periodogram.
\item \code{ts_length_considered} - the time series length used for the computation.
\item \code{expl_var} -  The explained variance of the frequency component `freq_AVP.
}

a named list containing
(i) the best-fitting frequency estimated with the Adapted Variable Period Technique,
(ii) the corresponding maximum value of the adapted periodogram,
(iii) the time series length used for the computation and
(iv) the explained variance.
}
\description{
This function computes the frequency of a time series based on
the adapted variable period technique (see: ...).
In contrast to the direct evaluation of the power spectrum,
this method is considering cycle lengths which correspond to an
integer multiple of the sampling frequency.
}
\examples{
# Let's consider three synthetic time series and an example from the attached data set:
x1 <- sin( 0.05 * 2 * pi * (1:100))
x2 <- sin( 0.05 * 2 * pi * (1:100)) + 1.1 * sin(0.07 * 2 * pi * (1:100))
x3 <- sin( 0.05 * 2 * pi * (1:100)) + sin(0.07 * 2 * pi * (1:100))
ecg_6  <- MiceFibECGs[,6]
freq_adapted_variable_period (x1)
freq_adapted_variable_period(x2)
freq_adapted_variable_period (x3)

freq_adapted_variable_period(ecg_6, delta_t = 0.001)
# which results in a different frequency and in particular
# in a higher explained variance than the standard spectral method
freq_argmax_periodogram(ecg_6, delta_t = 0.001)

}
\references{
{ Diaz-Maue L, Witt A, Nobach H: \emph{Unraveling Cardiac Arrhythmia Frequency, Comparative Analysis Using Time and Frequency Domain Algorithms}. submitted to Frontiers in Signal Processing (2025)}
}
