% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_argmax_periodogram.R
\name{freq_argmax_periodogram}
\alias{freq_argmax_periodogram}
\title{Argmax of the Periodogram}
\usage{
freq_argmax_periodogram(
  x,
  delta_t = 1,
  f_min = 0,
  f_max = Inf,
  welch_window = FALSE
)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{f_min}{lower endpoint of the frequency interval}

\item{f_max}{upper endpoint of the frequency interval}

\item{welch_window}{if TRUE, the time series will be multiplied with a Welch window before Fourier transform}
}
\value{
a list containing the argmax frequency and the explained variance
(i.e., the fraction of variance explained by this specific frequency), and the used endpoints of the frequency interval
}
\description{
This function computes the argmax of the periodogram (with respect to a given frequency interval) for a time series,
i.e., the frequency which corresponds to the maximum peak in the periodogram in the considered frequency range.
This frequency is also known as dominant frequency.
}
\examples{
x1 = sin(0.11*2*pi*(1:100));
x2 = sin(0.11*2*pi*(1:100))+ 1.01*sin(0.17*2*pi*(1:100));
x3 = sin(0.10*2*pi*(1:100))+ sin(0.20*2*pi*(1:100));
argmax_periodogram(x1)
argmax_periodogram(x2)
argmax_periodogram(x2, min_freq=0.05, max_freq = 0.15)
argmax_periodogram(x3)

}
