% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_argmax_periodogram.R
\name{freq_argmax_periodogram}
\alias{freq_argmax_periodogram}
\title{Argmax of the Periodogram}
\usage{
freq_argmax_periodogram(
  x,
  delta_t = 1,
  f_min = 0,
  f_max = Inf,
  welch_window = FALSE
)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{f_min}{lower endpoint of the frequency interval}

\item{f_max}{upper endpoint of the frequency interval}

\item{welch_window}{if TRUE, the time series will be multiplied with a Welch window before Fourier transform}
}
\value{
a named list containing the following components:
\itemize{
\item \code{freq_argmax_periodogram} - the best-fitting frequency estimated with the Adapted Variable Period Technique.
\item \code{Max_P} - the corresponding maximum value of the periodogram.
\item \code{expl_var} -  the explained variance of the frequency component \code{freq_argmax_periodogram}.
}

a named list containing
(i) the argmax frequency,
(ii) the maximum (i.e., the Periodogram at argmax) and
(iii) the explained variance (i.e., the fraction of variance explained by the argmax frequency).
}
\description{
This function computes the argmax of the periodogram (with respect to a given frequency interval) for a time series,
i.e., the frequency which corresponds to the maximum peak in the periodogram in the considered frequency range.
This frequency is also known as fundamental or dominant frequency.
}
\examples{
# Let's consider two synthetic time series and an example from the attached data set:
x1 <- sin( 0.05 * 2 * pi * (1:100))
x2 <- sin( 0.05 * 2 * pi * (1:100)) + 1.01 * sin(0.07 * 2 * pi * (1:100))
ecg_6  <- MiceFibECGs[,6]

freq_argmax_periodogram(x1)
# Welch windowing can change the expected variance
freq_argmax_periodogram(x1, welch_window = TRUE)

freq_argmax_periodogram(x2)
# The result can depend on the considered frequency
freq_argmax_periodogram(x2, f_min = 0.01, f_max = 0.06)

freq_argmax_periodogram(ecg_6, delta_t = 0.001)

}
\references{
{ Diaz-Maue L, Witt A, Nobach H: \emph{Unraveling Cardiac Arrhythmia Frequency, Comparative Analysis Using Time and Frequency Domain Algorithms}. submitted to Frontiers in Signal Processing (2025)}
}
