% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_fitted_sinusoidals.R
\name{freq_fitted_sinusoidals}
\alias{freq_fitted_sinusoidals}
\title{Frequency Estimation Using Fitted Sinusoidal Functions for a Set of Frequencies}
\usage{
freq_fitted_sinusoidals(x, delta_t = 1, test_freqs)
}
\arguments{
\item{x}{a real valued array containing the values of the time series}

\item{delta_t}{a positive real value, the sampling time, default value is delta_t = 1.0}

\item{test_freqs}{a real-valued vector with positive values, the test frequencies}
}
\value{
a named list containing the following components:
\itemize{
\item \code{freq_fitted} - the best-fitting frequency.
\item \code{expl_var} -  the explained variance of the frequency component \code{freq_fitted}.
\item \code{params} - the model parameters for offset + amp * sin (2\emph{pi}freq + phase).
\item \code{model_ts} -  the fitted model.
}
}
\description{
This function is determining the fundamental frequency by
identifying the best fitting sinusoidal function with respect to a given
range of frequencies. It identifies the frequency that corresponds to
the maximum explained variance.
}
\examples{
# Let's consider three synthetic time series and an example from the attached data set:
x1 <- sin( 0.05 * 2 * pi * (1 : 100))
x2 <- x1 + 1.01 * sin( 0.07 * 2 * pi * (1:100))
x3 <- x1 + 1.0 * sin( 0.07 * 2 * pi * (1:100))
ecg_6  <- MiceFibECGs[,6]

freq_fitted_sinusoidals(x1, test_freqs = seq(0.03, 0.20, by = 0.01))
freq_fitted_sinusoidals(x2, test_freqs = seq(0.03, 0.20, by = 0.01))
# The result can depend on the considered set of frequencies
freq_fitted_sinusoidals(x2, test_freqs = seq(0.02, 0.06, by = 0.01))
freq_fitted_sinusoidals(x2,  test_freqs = runif(100))
freq_fitted_sinusoidals(x3,  test_freqs = seq(0.03, 0.20, by = 0.01))

res <- freq_fitted_sinusoidals(ecg_6,  test_freqs = 1:30, delta_t = 0.001)
print(c("freq_fitted:",res$freq_fitted))
print(c("expl_var:",round(res$expl_var, 2)))
# Let's plot the fitted model

xlab = "Time [ms]"
ylab = "ECG"
plot(1:1000, ecg_6, xlab = xlab, ylab = ylab, "l", lwd = 1.5)
lines(1:1000, res$model, col = "red", lty = 2)
}
\references{
{ Diaz-Maue L, Witt A, Nobach H: \emph{Unraveling Cardiac Arrhythmia Frequency, Comparative Analysis Using Time and Frequency Domain Algorithms}. submitted to Frontiers in Signal Processing (2025)}
}
