% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_Lorentz_fit.R
\name{freq_Lorentz_fit}
\alias{freq_Lorentz_fit}
\title{Frequency Estimation Using a Lorentz Fit to the Periodogram}
\usage{
freq_Lorentz_fit(
  x,
  delta_t = 1,
  welch_window = FALSE,
  f_min = 0,
  f_max = Inf,
  algorithm = "Whittle"
)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{welch_window}{if TRUE, the time series will be multiplied with a Welch window before Fourier transform}
}
\value{
a list containing the parameters of the Lorentz fit.
}
\description{
This function is fitting a Lorentz function to the periodogram of the time series x.
}
\examples{
x1 = sin(0.11*2*pi*(1:100));
x2 = sin(0.11*2*pi*(1:100))+ 1.01*sin(0.17*2*pi*(1:100));
x3 = sin(0.10*2*pi*(1:100))+ sin(0.20*2*pi*(1:100));
freq_mode_periodogram(x1)
freq_mode_periodogram(x2)
freq_mode_periodogram(x2, min_freq=0.05, max_freq = 0.15)
freq_mode_periodogram(x3)
}
