% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_max2max.R
\name{freq_max2max}
\alias{freq_max2max}
\title{Frequency Estimation by Evaluating the Max-to-Max Time Intervals}
\usage{
freq_max2max(x, delta_t = 1, ..., .warn = TRUE)
}
\arguments{
\item{x}{a non-infinite real-valued numeric vector containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{...}{optional arguments to be passed to the function splus2R::peaks. Otherwise span = 11, strict = TRUE and endbehavior = 0 are used.}

\item{.warn}{Logical; if TRUE (default), messages are shown for edge cases.}
}
\value{
a named list containing the following components:
\itemize{
\item \code{freq_max2max} - the estimated frequency.
\item \code{c_v} - the coefficient of variation of the inter-maxima-intervals.
\item \code{n_max} - the number of identified local maxima.
\item \code{maxima_sampled} -   the time series indices of the identified local maxima.
}
}
\description{
Frequency estimation based on the determination of the average peak-to-peak distance.
This function requires smooth and oscillating time series.
The peaks are identified with the splus2R::peaks function.
}
\examples{
# Let's consider two synthetic time series and an example of the attached data set
x1 <- sin( 0.05 * 2 * pi * (1:100))
x2 <- sin(0.02 *  2*pi * (1:100)) + 1.1 * sin( 0.025 * 2*pi * (1:100))

ecg_6  <- MiceFibECGs[,6]

freq_max2max(x1)
freq_max2max(x2)

freq_max2max(ecg_6, delta_t = 0.001)
# Changing optional parameters of splus2R::peaks can change the result
freq_max2max(ecg_6, delta_t = 0.001, span = 35)
}
\references{
{ Diaz-Maue L, Witt A, Nobach H: \emph{Unraveling Cardiac Arrhythmia Frequency, Comparative Analysis Using Time and Frequency Domain Algorithms}. Front. Signal Process., Sec. Biomed. Signal Process. 5 (2025),  \url{https://doi.org/10.3389/frsip.2025.1707422}}
}
