% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_max2max.R
\name{freq_max2max}
\alias{freq_max2max}
\title{Frequency Estimation by Evaluating the Max-to-Max Time Intervals}
\usage{
freq_max2max(x, delta_t = 1, ...)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{...}{arguments to be passed to the function splus2R::peaks. Here span = 11, strict = TRUE and endbehavior = 0 are used as default values.}
}
\value{
a vector with the estimated frequency, its standard deviation and its interquartile range
}
\description{
Frequency Estimation by Evaluating the Max-to-Max Time Intervals
}
\examples{
x1 = sin(0.11*2*pi*(1:100));
x2 = sin(0.11*2*pi*(1:100))+ 1.01*sin(0.17*2*pi*(1:100));
freq_max2max(x1)
freq_max2max(x2)
freq_max2max(sunspots, delta_t=1./12, span = 51)
}
