% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_threshold_crossing.R
\name{freq_threshold_crossing}
\alias{freq_threshold_crossing}
\title{Frequency Estimation by Evaluating the Time Intervals between the arguments of succeeding local maxima}
\usage{
freq_threshold_crossing(x, delta_t = 1, thresh, ...)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}
}
\value{
a list with the three elements, estimated frequency, the standard deviation of the estimated frequency, and
the interquartile range of the estimated frequency.
}
\description{
Frequency Estimation by Evaluating the Time Intervals between the arguments of succeeding local maxima
}
\examples{
x1 = sin(0.11*2*pi*(1:100));
x2 = sin(0.11*2*pi*(1:100))+ 1.01*sin(0.17*2*pi*(1:100));
freq_threshold_crossing(x1)
freq_threshold_crossing(x2)
}
