% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_threshold_crossing.R
\name{freq_threshold_crossing}
\alias{freq_threshold_crossing}
\title{Frequency Estimation Using the Interval Length between Succeeding Upward Threshold Crossings}
\usage{
freq_threshold_crossing(x, delta_t = 1, thresh = 0)
}
\arguments{
\item{x}{x a non-infinite real-valued numeric vector containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{thresh}{a real value, the considered threshold,  default value is thresh = 0.0}
}
\value{
a named list containing the following components:
\itemize{
\item \code{freq_threshold_crossing} - the estimated frequency.
\item \code{c_v} - the coefficient of variation of the time intervals between succeeding upward threshold crossings.
\item \code{n_thresh} - the number of upward threshold crossings.
\item \code{threshold_crossings} -  the time series indices of upward threshold crossings.
}
}
\description{
This frequency estimation function computes the fundamental frequency as the inverse of the  average distance
between successive upward threshold crossings. It requires smooth and oscillating time series.
}
\examples{
# Let's consider a synthetic time series and an example from the attached data set:
x1 <- sin( 0.05 * 2*pi * (1:100))
ecg_6  <- MiceFibECGs[,6]

freq_threshold_crossing(x1)

# Changing the threshold can change the resulting frequency
freq_threshold_crossing(x1, thresh = 0.5)

# The irregularity of the fibrillation is indicated by a high value of the coefficient of variation
freq_threshold_crossing(ecg_6, delta_t = 0.001)

}
\references{
{ Diaz-Maue L, Witt A, Nobach H: \emph{Unraveling Cardiac Arrhythmia Frequency, Comparative Analysis Using Time and Frequency Domain Algorithms}. submitted to Frontiers in Signal Processing (2025)}
}
