% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{Periodogram}
\alias{Periodogram}
\title{Periodogram of a Real-Valued Time Series}
\usage{
Periodogram(x, delta_t = 1, welch_window = FALSE)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a real value, the sampling time, default value is delta_t = 1.0}

\item{welch_window}{if TRUE, the time series will be multiplied with a Welch window before Fourier transform}
}
\value{
a list with the elements Frequency and Power
}
\description{
Periodogram of a Real-Valued Time Series
}
\examples{
x1 <- rnorm(32)
Periodogram(x1)
Periodogram(x1, delta = 0.25)
Periodogram(x1, welch_window = FALSE)
x2 <- sin(0.1 * 2 * pi * (1:100));
plot(Periodogram(x2), type = 'l', co="orange")
points(Periodogram(x2), pch=21, cex=0.2)

}
