% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{Periodogram}
\alias{Periodogram}
\title{Periodogram of a Real-Valued Time Series}
\usage{
Periodogram(x, delta_t = 1, welch_window = FALSE)
}
\arguments{
\item{x}{real valued array containing the values of the time series}

\item{delta_t}{a positive real value, the sampling time, default value is delta_t = 1.0}

\item{welch_window}{if TRUE, the time series will be multiplied with a Welch window before Fourier transform}
}
\value{
a named list containing the following components:
\itemize{
\item \code{Freqeuncy} -  a real valued array  containing the frequencies.
\item \code{PSD} -  a real valued array containing the computed power spectral density.
}
}
\description{
This function computes the Periodogram, i.e., a simple estimator of the power spectral density of a real valued time series.
}
\examples{
x1 <- rnorm(256)
P1 <- Periodogram(x1, delta = 0.25)
P1_w <-Periodogram(x1, delta_t = 0.25, welch_window = TRUE)
x2 <- sin(0.1 * 2 * pi * (1:100));
P2 <- Periodogram(x2)

plot( P1$Frequency, P1$Power, type = 'l', col = "orange")
points(P1$Frequency, P1$Power, pch = 21, , bg = "black", cex = 0.5)
lines(P1_w$Frequency, P1_w$Power, cex = 0.2, col="violet")
points(P1_w$Frequency, P1_w$Power, pch = 25, bg = "blue", col="blue", cex = 0.5)

plot( P2$Frequency, P2$Power, type = 'l', col = "orange")
points(P2$Frequency, P2$Power, pch = 21, bg = "black", cex = 0.2)

}
\references{
{ Priestley, MB: \emph{Spectral Analysis and Time series}. Academic Press, London (1981)}
}
